+++ lib/gamefic/command.rb
@@ -1,18 +1,17 @@
 # frozen_string_literal: true
 
 module Gamefic
-  # A decomposition of a text-based command into its verb and arguments.
-  #
-  # Commands are typically derived from tokenization against syntaxes.
+  # A concrete representation of an input as a verb and an array of arguments.
   #
   class Command
     # @return [Symbol]
     attr_reader :verb
 
-    # @return [Array<String>]
+    # @return [Array<Array<Entity>, Entity, String>]
     attr_reader :arguments
-    alias tokens arguments
 
+    # @param verb [Symbol]
+    # @param arguments [Array<Array<Entity>, Entity, String>]
     def initialize verb, arguments
       @verb = verb
       @arguments = arguments