+++ spec/gamefic/syntax_spec.rb
@@ -23,23 +23,23 @@ describe Gamefic::Syntax do
 
   it "tokenizes complex phrases" do
     syntax = Gamefic::Syntax.new "get :foo from :bar", "get :foo :bar"
-    command = syntax.tokenize("get the one from the two")
-    expect(command.verb).to be(:get)
-    expect(command.arguments).to eq(['the one', 'the two'])
+    expression = syntax.tokenize("get the one from the two")
+    expect(expression.verb).to be(:get)
+    expect(expression.tokens).to eq(['the one', 'the two'])
   end
 
   it "tokenizes verbs with multiple words" do
     syntax = Gamefic::Syntax.new "put up :thing", "put_up :thing"
-    command = syntax.tokenize("put up your dukes")
-    expect(command.verb).to be(:put_up)
-    expect(command.arguments).to eq(['your dukes'])
+    expression = syntax.tokenize("put up your dukes")
+    expect(expression.verb).to be(:put_up)
+    expect(expression.tokens).to eq(['your dukes'])
   end
 
-  it "tokenizes arguments with multiple words" do
+  it "tokenizes tokens with multiple words" do
     syntax = Gamefic::Syntax.new "original :foo", "converted :foo"
-    command = syntax.tokenize("original one and two")
-    expect(command.verb).to be(:converted)
-    expect(command.arguments).to eq(['one and two'])
+    expression = syntax.tokenize("original one and two")
+    expect(expression.verb).to be(:converted)
+    expect(expression.tokens).to eq(['one and two'])
   end
 
   it "does case-insensitive matches" do
@@ -51,16 +51,16 @@ describe Gamefic::Syntax do
 
   it "tokenizes syntaxes that switch from opening argument to command" do
     syntax = Gamefic::Syntax.new ":vendor make :product", "order :vendor :product"
-    command = syntax.tokenize('bob make cookies')
-    expect(command.verb).to eq(:order)
-    expect(command.arguments).to eq(['bob', 'cookies'])
+    expression = syntax.tokenize('bob make cookies')
+    expect(expression.verb).to eq(:order)
+    expect(expression.tokens).to eq(['bob', 'cookies'])
   end
 
   it "tokenizes syntaxes that have commands without opening verbs" do
     syntax = Gamefic::Syntax.new ":vendor make :product", ":vendor manufacture :product"
-    command = syntax.tokenize('bob make cookies')
-    expect(command.verb).to eq(nil)
-    expect(command.arguments).to eq(['bob', 'manufacture', 'cookies'])
+    expression = syntax.tokenize('bob make cookies')
+    expect(expression.verb).to eq(nil)
+    expect(expression.tokens).to eq(['bob', 'manufacture', 'cookies'])
   end
 
   it 'equals other based on signature' do
@@ -86,10 +86,10 @@ describe Gamefic::Syntax do
     s1 = Gamefic::Syntax.new 'make :a from :b', 'make :a :b'
     s4 = Gamefic::Syntax.new 'use :x :y', 'use :x :y'
     s3 = Gamefic::Syntax.new 'use :x on :y', 'use :x :y'
-    commands = Gamefic::Syntax.tokenize 'make thing from material', [s1, s2, s3, s4]
-    expect(commands.map(&:verb)).to eq([:make, :make])
-    expect(commands.first.arguments).to eq(['thing', 'material'])
-    expect(commands.last.arguments).to eq(['thing from material', nil])
+    expressions = Gamefic::Syntax.tokenize 'make thing from material', [s1, s2, s3, s4]
+    expect(expressions.map(&:verb)).to eq([:make, :make])
+    expect(expressions.first.tokens).to eq(['thing', 'material'])
+    expect(expressions.last.tokens).to eq(['thing from material', nil])
   end
 
   it 'validates command parameters' do