+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.1.0
 - Dispatcher prioritizes strict token matches
+- Tokenize expressions and execute commands
 
 ## 3.0.0 - January 27, 2024
 - Instantiate subplots from snapshots